minor bug fixes

+++ src/random_uniform_generator.cpp
@@ -1,5 +1,3 @@
-//#include "../include/random_uniform_generator.hpp"
-
 #include <random>
 
 class RandomNumberBetween
@@ -12,7 +10,7 @@ public:
 	}	
 	int RandomNumberBetween::operator()()
 	{
-		return distribution_(random_engine_);
+		return static_cast<unsigned int>(distribution_(random_engine_));
 	}
 private:
 	std::mt19937 random_engine_;
